+++ services/cost_management/cost_management_service/src/main/java/org/changppo/cost_management_service/service/member/oauth/OAuthProperties.java
@@ -8,7 +8,12 @@ import org.springframework.context.annotation.Configuration;
 @Getter
 @Setter
 @Configuration
-@ConfigurationProperties(prefix = "adminkey")
-public class AdminKeyProperties {
-    private String kakao;
+@ConfigurationProperties(prefix = "oauth")
+public class OAuthProperties {
+    private Kakao kakao;
+    @Getter
+    @Setter
+    public static class Kakao {
+        private String adminKey;
+    }
 }
